temporalguard-sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "temporalguard"
version = "0.1.0"
description = "Runtime verification for LLM conversations using past-time temporal logic"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "openai>=2.29.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.110"]
langchain = ["langchain-core>=0.2"]
openai = ["openai>=1.0"]
all = ["fastapi>=0.110", "langchain-core>=0.2", "openai>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/temporalguard"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (Docker/Ollama)",
]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.20",
]
